% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{debugMessage}
\alias{debugMessage}
\title{Print Debug Message}
\usage{
debugMessage(message, verbose = FALSE)
}
\arguments{
\item{message}{A character string representing the message to be printed.}

\item{verbose}{A logical value indicating whether the verbose mode is
enabled. Defaults to \code{FALSE}.}
}
\value{
This function returns void.
}
\description{
This function is used internally to print debug messages when the verbose
mode is enabled.
}
\keyword{internal}
