% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPar.R
\name{doPar}
\alias{doPar}
\title{\verb{\%dopar\%} Implementation For \code{parabar} Backends}
\usage{
doPar(obj, expr, envir, data)
}
\value{
A list containing the results of the parallel computation.
}
\description{
See the documentation of \code{\link[foreach:foreach-ext]{foreach::foreach-ext}} for more information on how
to provide an implementation for the \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator.
}
\details{
Currently, the following \code{\link[foreach:foreach]{foreach::foreach}} arguments are supported:\tabular{rc}{
   Argument \tab Status \cr
   \code{...} \tab ✅ \cr
   \code{.combine} \tab ✅ \cr
   \code{.init} \tab ✅ \cr
   \code{.final} \tab ✅ \cr
   \code{.inorder} \tab ✅ \cr
   \code{.multicombine} \tab ✅ \cr
   \code{.maxcombine} \tab ✅ \cr
   \code{.errorhandling} \tab ✅ \cr
   \code{.packages} \tab ✅ \cr
   \code{.export} \tab ✅ \cr
   \code{.noexport} \tab ✅ \cr
   \code{.verbose} \tab ✅ \cr
}


Where:
\itemize{
\item ✅ indicates that the argument is fully supported
\item ❌ indicates that the argument is not supported
\item ❗️ indicates that the argument is partially supported
\item ❓ indicates that I am unsure about what the argument does
}
}
\keyword{internal}
