% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doParabar-package.R
\docType{package}
\name{doParabar-package}
\alias{doParabar}
\alias{doParabar-package}
\title{Foreach Parallel Adaptor For \code{parabar} Backends}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a `foreach` parallel adaptor for `parabar` backends.
}
\details{
The \code{\link{doParabar}} package acts as a \code{\link[foreach:foreach]{foreach::foreach}}
parallel adaptor for \code{\link[parabar:parabar-package]{parabar::parabar}} backends. It provides a minimal
implementation for the \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator, enabling seamless
integration of the \code{\link[parabar:parabar-package]{parabar::parabar}} package with the \code{\link[foreach:foreach]{foreach::foreach}}
package. For more details, please consult de documentation for
\code{\link[parabar:parabar-package]{parabar::parabar}} at
\href{https://parabar.mihaiconstantin.com}{parabar.mihaiconstantin.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mihaiconstantin/doParabar}
  \item Report bugs at \url{https://github.com/mihaiconstantin/doParabar/issues}
}

}
\author{
\strong{Maintainer}: Mihai Constantin \email{mihai@mihaiconstantin.com} (\href{https://orcid.org/0000-0002-6460-0107}{ORCID})

}
\keyword{internal}
