% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getNumberOfWorkers}
\alias{getNumberOfWorkers}
\title{Get the Number of Workers}
\usage{
getNumberOfWorkers(backend)
}
\arguments{
\item{backend}{An object of class \code{\link[parabar:Backend]{parabar::Backend}} representing the
backend used for the \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator implementation.}
}
\value{
An integer representing the number of workers for the registered
backend.
}
\description{
Get the Number of Workers
}
\keyword{internal}
