% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\docType{data}
\name{LOGO}
\alias{LOGO}
\title{The Package Logo}
\format{
An object of class \code{character} containing the \code{ASCII} logo.
}
\usage{
LOGO
}
\description{
The logo is generated by \code{\link[parabar:make_logo]{parabar::make_logo()}} and displayed on package
attach for interactive \code{R} sessions.
}
\details{
The logo is based on the \code{doParabar-logo.txt} template file located under the
\code{inst/assets} directory of the package.
}
\seealso{
\code{\link[parabar:make_logo]{parabar::make_logo()}}
}
\keyword{internal}
