% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parabarInfo}
\alias{parabarInfo}
\title{Registered Implementation Information}
\usage{
parabarInfo(data, item)
}
\arguments{
\item{data}{An object of class \code{\link[parabar:Backend]{parabar::Backend}} representing the backend
used for the \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator implementation.}

\item{item}{A character string specifying the information item to be
retrieved about the registered \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator implementation.}
}
\value{
The requested information item.
}
\description{
This function is used internally by \code{\link[foreach:foreach]{foreach::foreach}} to retrieve
information about the registered \code{\link[foreach:foreach]{foreach::\%dopar\%}} operator
implementation. More specifically, this function is called by
\code{\link[foreach:getDoParWorkers]{foreach::getDoParName()}}, \code{\link[foreach:getDoParWorkers]{foreach::getDoParRegistered()}},
\code{\link[foreach:getDoParWorkers]{foreach::getDoParVersion()}}, and \code{\link[foreach:getDoParWorkers]{foreach::getDoParWorkers()}}.
}
\keyword{internal}
